JailbreakFile/requirements.txt
groq>=0.9.0
PyYAML>=6.0
python-dotenv>=1.0.1